+++ rollup.config.js
@@ -2,7 +2,7 @@ import commonjs from 'rollup-plugin-commonjs';
 import nodeResolve from 'rollup-plugin-node-resolve';
 
 export default {
-	entry: 'index.js',
+	entry: 'config/.build_tmp.js',
 	dest: 'dist/driver-lib.js',
 	moduleName: '',
 	context: 'window',
@@ -21,7 +21,8 @@ export default {
 			sourceMap: false,
 			include: [
 				'node_modules/hogan.js/**',
-				'node_modules/url-search-params/**'
+				'node_modules/url-search-params/**',
+				'node_modules/lodash/**'
 			]
 		})
 	]