+++ CHANGELOG.md
@@ -1,2 +1,3 @@
-- 0.1.1 Use babel-polyfill instead of core-js
-- 0.1.0 init
+- 0.1.2 Build lodash from config file. Reduce size of dist min file from 332KB to 292KB.
+- 0.1.1 Use babel-polyfill instead of core-js.
+- 0.1.0 init.